requirements.txt
altgraph==0.17.5
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Flask==3.0.0
idna==3.11
invoke==2.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
paramiko==4.0.0
pefile==2024.8.26
Pillow==10.1.0
psutil==5.9.6
pycparser==3.0
pycryptodome==3.19.0
pycryptodomex==3.23.0
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
PyNaCl==1.6.2
pyspnego==0.12.1
pywin32==306
pywin32-ctypes==0.2.3
requests==2.31.0
setuptools==82.0.1
smbprotocol==1.16.0
sspilib==0.5.0
urllib3==2.6.3
Werkzeug==3.1.6
wheel==0.46.3
WMI==1.5.1
